refactor(core): split order item resolution modules
- move single item validation and pricing into orderItemResolution
- move cached batch request resolution and quote composition into focused modules
- keep orderItems as a compatibility export surface while removing its hotspot logic
